Move Physicians page components into /physicians
