+++ app/physicians/PhysicianComponents/PhysicianSearchArea.tsx
@@ -5,7 +5,7 @@ import Fuse from "fuse.js";
 // TYPES
 import { PhysicianBioType } from "@/app/constants/types";
 import SearchBar from "@/app/components/ui/SearchBar";
-import PhysiciansGridLayout from "@/app/components/PhysicianComponents/PhysiciansGridLayoutSection";
+import PhysiciansGridLayout from "@/app/physicians/PhysicianComponents/PhysiciansGridLayoutSection";
 
 interface SortedPhysicians {
   mdPhysicians: PhysicianBioType[];
