+++ app/physicians/page.tsx
@@ -2,7 +2,7 @@
 import { getPhysicians } from "@/app/utils/contentful";
 // TYPES
 import { PhysicianBioType } from "@/app/constants/types";
-import PhysicianSearchArea from "@/app/components/PhysicianComponents/PhysicianSearchArea";
+import PhysicianSearchArea from "@/app/physicians/PhysicianComponents/PhysicianSearchArea";
 
 export default async function Physicians() {
   const physicians = (await getPhysicians()) as unknown as PhysicianBioType[];
